Add tests
Also extracts parseCommand() to a separate file for simpler testing
setup.
