+++ src/main.cpp
@@ -33,6 +33,7 @@
 
 #include "mainwindow.h"
 #include "qterminalapp.h"
+#include "qterminalutils.h"
 #include "terminalconfig.h"
 
 #define out
@@ -70,49 +71,6 @@ QTerminalApp * QTerminalApp::m_instance = nullptr;
 {
     printf("%s\n", QTERMINAL_VERSION);
     exit(code);
-}
-
-QStringList parse_command(const QString& str)
-{
-    const QRegularExpression separator(QString::fromLatin1(R"('|(?<!\\)(\\{2})*(\s|")|\z)"));
-    const QRegularExpression doubleQuote(QString::fromLatin1(R"((?<!\\)(\\{2})*")"));
-    const QRegularExpression escapedSpace(QString::fromLatin1(R"(\\(\\{2})*\s)"));
-    const QRegularExpression singleQuote(QStringLiteral("'"));
-
-    QStringList list;
-    QRegularExpressionMatch match;
-    int index = 0;
-    int nextIndex;
-    while((nextIndex = str.indexOf(separator, index, &match)) != -1)
-    {
-        if (nextIndex > index)
-        {
-            list << str.mid(index, nextIndex - index).replace(escapedSpace, QStringLiteral(" "));
-        }
-        if (match.capturedLength() == 0)
-        { // end of string ("\z") is matched
-            break;
-        }
-        index = nextIndex + match.capturedLength();
-        auto c = str.at(index - 1); // last matched character
-        if (!c.isSpace())
-        { // a single quote or an unescaped double quote is matched
-            nextIndex = str.indexOf(c == QLatin1Char('\'') ? singleQuote : doubleQuote, index, &match);
-            if (nextIndex == -1)
-            { // the quote is not closed
-                break;
-            }
-            else
-            {
-                if (nextIndex > index)
-                {
-                    list << str.mid(index, nextIndex - index).replace(escapedSpace, QStringLiteral(" "));
-                }
-                index = nextIndex + match.capturedLength();
-            }
-        }
-    }
-    return list;
 }
 
 void parse_args(int argc, char* argv[], QString& workdir, QStringList & shell_command, out bool& dropMode)
