+++ test/qterminal_test.h
@@ -1,0 +1,32 @@
+/***************************************************************************
+ *   Copyright (C) 2022 by LXQt team                                       *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ ***************************************************************************/
+
+#ifndef QTERMINAL_TEST_H
+#define QTERMINAL_TEST_H
+
+#include <QObject>
+
+class QTerminalTest : public QObject
+{
+    Q_OBJECT
+
+    // Each private slot is a test function
+private Q_SLOTS:
+    void testParseCommand();
+};
+
+#endif
